Added BlogPostsListActivityInstrumentedTest
