+++ app/src/androidTest/java/com/coveros/coverosmobileapp/website/InstrumentedWebViewTest.java
@@ -2,6 +2,7 @@ package com.coveros.coverosmobileapp.website;
 
 import android.app.AlertDialog;
 import android.content.Intent;
+import android.support.test.annotation.UiThreadTest;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.rule.ServiceTestRule;
 import android.support.test.runner.AndroidJUnit4;
@@ -86,12 +87,14 @@ public class InstrumentedWebViewTest extends LooperTestSuite {
     //which means the creation of Buttons/TextViews in xml activity
     //linking the buttons and strings together
     @Test
+    @UiThreadTest
     public void slideNavigationMenuOpen(){
         DrawerLayout menu = (DrawerLayout) mMainActivity.getActivity().findViewById(R.id.drawer_layout);
         menu.openDrawer(Gravity.START);
         assertTrue(menu.isDrawerOpen(Gravity.START));
     }
     @Test
+    @UiThreadTest
     public void slideNavigationMenuClosed(){
         DrawerLayout menu = (DrawerLayout) mMainActivity.getActivity().findViewById(R.id.drawer_layout);
         menu.openDrawer(Gravity.START);
